State before ripping out variable inheritance
Trying to get variable inheritance to work, in standard Ansible fashion.
The problem this committer is facing is the "double moustache" and the
"correct" way of using hostvars[inventory_hostname] won't properly
inherit from group_vars.

This committer has dedicated to much time to this problem. It's past the
point of diminishing returns when a simpler, more reliable method
exists.

Moving to passing any default overrides as command line --extra-vars